Map listing helpers are in their own file
listings_helper was a bit tricky to navigate.
